Fix segfault in `removeVertexNode` due to treating the modulo as unsigned
The hash in `_hashVertex` produces out of range values if `(vertex->lat + vertex->lon)`
was negative, because it would produce a negative module (such as -1) and then cast it to
an unsigned integer type. This index would then cause a segfault.

testH3SetToLinkedGeo.c is also updated to not `calloc` a fixed size of 1.