Clarify that the reserved bits must be 0 for cell (mode 1) indexes.
Clarify the documentation and validation functions that the reserved
bits must be 0 for a cell index to be valid. This is done to prevent
valid indexes being different but refering to the same cell.

The high bit of an H3Index is also treated as a reserved bit
(previously documentation was silent on it.)

See #297.