FIX: Fix polyfill bug when vertex latitude exactly matches cell center (#603)
Fixes #595

The issue here is that rays cast directly through a vertex of the polygon were being counted twice, once for each adjoining segment. This uses the same "bias in an arbitrary direction" approach we used for longitude matches, which should also ensure no overlap between contiguous polygons, which I think was previously an issue.