Fix case where parent bounding box contains hole (#806)
This issue came up when comparing old/new output for countries: When the parent bounding box is fully contained by the outer polygon but fully contains a hole, we were still including the parent in the output. The fix is to check whether the bounding box contains a vertex of the hole, and fail fast if it does.